requirements.txt
Flask==2.0.3
pycryptodome==3.14.1
requests==2.27.1
